requirements.txt
Django==2.2.28
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
wincertstore==0.2
django-registration-redux==2.2